refactor: consolidate JSON request validation into a reusable function
