+++ internal/api/handlers/auth.go
@@ -5,7 +5,6 @@ import (
 	"encoding/json"
 	"errors"
 	"fmt"
-	"io"
 	"log/slog"
 	"net/http"
 
@@ -28,42 +27,8 @@ type authResponse struct {
 
 func NewAuthRegisterHandler(svc AuthService) http.Handler {
 	return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
-		if r.Header.Get("Content-Type") != "application/json" {
-			slog.Error(
-				"request with an empty or unsupported content type",
-				slog.String("content_type", r.Header.Get("Content-Type")),
-			)
-			http.Error(w, "wrong content type", http.StatusUnsupportedMediaType)
-			return
-		}
-
-		r.Body = http.MaxBytesReader(w, r.Body, 1<<20)
-		defer func() { _ = r.Body.Close() }()
-
-		dec := json.NewDecoder(r.Body)
-		dec.DisallowUnknownFields()
-
 		var authReq authRequest
-		if err := dec.Decode(&authReq); err != nil {
-			var mberr *http.MaxBytesError
-			slog.Warn("invalid JSON", slog.Any("error", err))
-			if errors.As(err, &mberr) {
-				http.Error(
-					w,
-					"request body too large",
-					http.StatusRequestEntityTooLarge,
-				)
-				return
-			}
-			http.Error(w, "invalid JSON", http.StatusBadRequest)
-			return
-		}
-
-		if err := dec.Decode(&struct{}{}); err != io.EOF {
-			slog.Warn("invalid JSON", slog.Any("error", err))
-			http.Error(w, "invalid JSON", http.StatusBadRequest)
-			return
-		}
+		ValidateParseJSONRequest(w, r, &authReq)
 
 		token, err := svc.Register(r.Context(), authReq.Login, authReq.Password)
 		if err != nil {
@@ -91,68 +56,14 @@ func NewAuthRegisterHandler(svc AuthService) http.Handler {
 			return
 		}
 
-		authResp := &authResponse{
-			Token: token,
-		}
-
-		b, err := json.Marshal(authResp)
-		if err != nil {
-			slog.Warn("failed to marshal json response", slog.Any("error", err))
-			http.Error(
-				w,
-				http.StatusText(http.StatusInternalServerError),
-				http.StatusInternalServerError,
-			)
-			return
-		}
-
-		w.Header().Set("Authorization", fmt.Sprintf("Bearer %s", token))
-		w.Header().Set("Content-Type", "application/json")
-		if _, err := w.Write(b); err != nil {
-			slog.Warn("failed to write response", slog.Any("error", err))
-			return
-		}
+		authJSONResponse(w, r, token)
 	})
 }
 
 func NewAuthLoginHandler(svc AuthService) http.Handler {
 	return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
-		if r.Header.Get("Content-Type") != "application/json" {
-			slog.Error(
-				"request with an empty or unsupported content type",
-				slog.String("content_type", r.Header.Get("Content-Type")),
-			)
-			http.Error(w, "wrong content type", http.StatusUnsupportedMediaType)
-			return
-		}
-
-		r.Body = http.MaxBytesReader(w, r.Body, 1<<20)
-		defer func() { _ = r.Body.Close() }()
-
-		dec := json.NewDecoder(r.Body)
-		dec.DisallowUnknownFields()
-
 		var authReq authRequest
-		if err := dec.Decode(&authReq); err != nil {
-			var mberr *http.MaxBytesError
-			slog.Warn("invalid JSON", slog.Any("error", err))
-			if errors.As(err, &mberr) {
-				http.Error(
-					w,
-					"request body too large",
-					http.StatusRequestEntityTooLarge,
-				)
-				return
-			}
-			http.Error(w, "invalid JSON", http.StatusBadRequest)
-			return
-		}
-
-		if err := dec.Decode(&struct{}{}); err != io.EOF {
-			slog.Warn("invalid JSON", slog.Any("error", err))
-			http.Error(w, "invalid JSON", http.StatusBadRequest)
-			return
-		}
+		ValidateParseJSONRequest(w, r, &authReq)
 
 		token, err := svc.Login(r.Context(), authReq.Login, authReq.Password)
 		if err != nil {
@@ -174,26 +85,34 @@ func NewAuthLoginHandler(svc AuthService) http.Handler {
 			return
 		}
 
-		authResp := &authResponse{
-			Token: token,
-		}
-
-		b, err := json.Marshal(authResp)
-		if err != nil {
-			slog.Warn("failed to marshal json response", slog.Any("error", err))
-			http.Error(
-				w,
-				http.StatusText(http.StatusInternalServerError),
-				http.StatusInternalServerError,
-			)
-			return
-		}
-
-		w.Header().Set("Authorization", fmt.Sprintf("Bearer %s", token))
-		w.Header().Set("Content-Type", "application/json")
-		if _, err := w.Write(b); err != nil {
-			slog.Warn("failed to write response", slog.Any("error", err))
-			return
-		}
+		authJSONResponse(w, r, token)
 	})
+}
+
+func authJSONResponse(
+	w http.ResponseWriter,
+	r *http.Request,
+	token string,
+) {
+	authResp := &authResponse{
+		Token: token,
+	}
+
+	b, err := json.Marshal(authResp)
+	if err != nil {
+		slog.Warn("failed to marshal json response", slog.Any("error", err))
+		http.Error(
+			w,
+			http.StatusText(http.StatusInternalServerError),
+			http.StatusInternalServerError,
+		)
+		return
+	}
+
+	w.Header().Set("Authorization", fmt.Sprintf("Bearer %s", authResp.Token))
+	w.Header().Set("Content-Type", "application/json")
+	if _, err := w.Write(b); err != nil {
+		slog.Warn("failed to write response", slog.Any("error", err))
+		return
+	}
 }
